Add files via upload
https://codeforces.com/problemset/problem/1195/B
Each evening after the dinner the SIS's students gather together to play the game of Sport Mafia.
For the tournament, Alya puts candies into the box, which will serve as a prize for a winner. To do that, she performs n actions. The first action performed is to put a single candy into the box. For each of the remaining moves she can choose from two options:
the first option, in case the box contains at least one candy, is to take exactly one candy out and eat it. This way the number of candies in the box decreased by 1;
the second option is to put candies in the box. In this case, Alya will put 1 more candy, than she put in the previous time.